requirements.txt
selenium==4.2.0
pytest==7.1.2
webdriver-manager==3.7.0
beautifulsoup4==4.11.1
python-dotenv==0.20.0
pandas==1.2.4